fix(#985): fix forest cover layer species average height from integer to float (#986)

+++ backend/src/main/java/ca/bc/gov/restapi/results/oracle/dto/cover/history/OpeningForestCoverHistoryDetailedSpeciesDto.java
@@ -16,7 +16,7 @@ public record OpeningForestCoverHistoryDetailedSpeciesDto(
     Long averageAge,
 
     @Schema(types = {"number", "null"}, requiredMode = Schema.RequiredMode.REQUIRED)
-    Long averageHeight
+    Float averageHeight
 ) {
 
 }
